rtc: pcf8523: properly handle oscillator stop bit
Upstream-commit: ede44c908d44b166a5b6bd7caacd105c2ff5a70f

The time and date register of the pcf8223 are undefined after a power
reset. Properly handle the OS bit and return -EINVAL when that bit is set.

It is properly removed when setting the time.

This solves an issue where the time and date may be valid for
rtc_valid_tm() but is not the current time.

Signed-off-by: Alexandre Belloni <alexandre.belloni@free-electrons.com>