dts: hb2: Add device-tree entries for canbus support
These entries have been missing since the dts rework add them
back in to make it easier to enable canbus support on the HB2
based boards.

Note that HDMI CEC shares lines with the CANBUS therefore the
hdmi_cec status needs to be set to disabled, and flexcan status
to okay.

Signed-off-by: Jon Nettleton <jon@solid-run.com>