MLK-17492 drm/imx: dec400d: set read config to '0x0' when bypass dec400d
When the DEC400D is set to bypass mode from decompressed mode,
the read config should be set to disable compression along with
the control register. Otherwise, the DEC400D cannot really leave
the decompressed mode. And the value '0x0' is suitable to be set
to read config register in this case.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
(cherry picked from commit 11d31a565dca268f64100735ac2a3ab112799171)