MLK-17648-1: drm: imx: dcss: add HDR10 module tables
This commit adds HDR10 tables as a header. Using a FW file is
problematic since the tables need to be available immediately after
boot. After the rootfs is mounted, as is the case for loading a FW file,
it's already too late if some conversion tables are needed.

This usually happens if the output pipe is configured as YUV420.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@nxp.com>
(cherry picked from commit dfb6fa2943119c2b371809f3b3b0463dc4ada3de)