mxc: ipu: Change order of stripe processing from up->down to down->up
When the output picture is too large to be handled the IPU hardware, it
is split into stripes that are processed independently by any of the IPUs
present in the SoC. There are some situations where the up (top) and the
down (bottom) stripe overlap by a few lines. While this is not a problem
for progressive frames, it may introduce a visible tearing line in the
middle of the screen when deinterlacing is in effect. Patch ENGR00161315
provides a workaround for this. However, due to recent changes in the
kernel memory management, the original ENGR00161315 will crash.
This patch removes ENGR00161315 and implements a more efficient solution
to this problem by changing the processing order of the stripes. Now the
down (bottom) stripe is scheduled first so that the up (top) stripe will
overwrite the area covered by both.

Signed-off-by: Rudi <r.ihle@s-t.de>