MXC video: don't reconfigure fb in mxc_hdmi_set_mode_to_vga_dvi
This is from an old Freescale commit which claimed that some displays
would not initialize properly unless they were first sent a vga
mode signal.  Sending the hdmi signal does not require reconfiguring
the underlying framebuffer, since we have no intention on drawing
anything thing to the screen in this state.

This also fixes the issue where booting with a cable unplugged and
then plugging in a cable results in a blank screen until something is
drawn on the screen.  This can lead people to believe that the
hdmi connection is not working when in fact it is.

Signed-off-by: Jon Nettleton <jon@solid-run.com>