gpu: galcore: remove locking from _QueryProcessPageTable
We are just looking up the physical address to a virtual address
that belongs to us.  I do not think that the locking is really
necessary here.  Since there are no other kernel drivers using
this locking and removing it fixes a kernel warning about slow
mutexes I will remove it.

Signed-off-by: Jon Nettleton <jon@solid-run.com>