ARM: dts: fsl-imx8mq-hb: Add specific graphics CMA region
We allocate two CMA regions now.  One in low memory that is the
default region and needs minimal space.  The second is a dedicated
graphics region that we pass into the drm and vivante drivers to use.

The logic behind this is that it allows the graphics memory to be
moved out of low memory as it isn't a requirement, and also limits
the devices that will be using the region.  This should minimize fragmentation
by other drivers requesting small blocks the break up larger memory regions
causing failures.

Signed-off-by: Jon Nettleton <jon@solid-run.com>